autoPy/requirements.txt
allure-pytest==2.8.22
jsonpath==0.82
loguru==0.5.1
pytest==6.1.2
PyYAML==5.3.1
requests==2.24.0
xlrd==1.2.0
yagmail==0.11.224
zipp==3.1.0
PyMySQL==0.10.1
